#54 : [FEAT] : Persistent bottomSheet로 변경
